# internal/routers/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a blog server restful api docs.",
        "title": "Blog Service API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/operation": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取多个操作日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operation"
                ],
                "summary": "获取多个操作日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "对象",
                        "name": "object",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.OperationLog"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "修改操作日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operation"
                ],
                "summary": "修改操作日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作日志ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改文章",
                        "name": "UpdateOperationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/app.CommResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "创建操作日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operation"
                ],
                "summary": "创建操作日志",
                "parameters": [
                    {
                        "description": "创建标签",
                        "name": "CreateOperationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/app.CommResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "删除操作日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operation"
                ],
                "summary": "删除操作日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作日志ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/app.CommResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/app.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.CommResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "ret_code": {
                    "type": "string"
                },
                "ret_desc": {
                    "type": "string"
                }
            }
        },
        "app.ErrResponse": {
            "type": "object",
            "properties": {
                "ret_code": {
                    "type": "string"
                },
                "ret_desc": {
                    "type": "string"
                }
            }
        },
        "app.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "ret_code": {
                    "type": "string"
                },
                "ret_desc": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.OperationLog": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "state": {
                    "type": "integer"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "service.CreateOperationRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "service.UpdateOperationRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}